Handle CONLL conversions where span is wider than contained entity
Label Studio allows an entity span to start or stop part-way through a token, or to include
leading or trailing whitespace.

This change ensures that:
* Leading/trailing whitespace in the span is ignored
* Any text outside of an entity span is tokenized separately.

For example, suppose the text says something like "Fred (1992) did something" and the user
has tagged 1992 as a date, but without the parentheses. This will be tokenized as
['Fred','(','1992',')']